code Structure updated along some ui changes
