+++ frontends/src/App.jsx
@@ -1,20 +1,32 @@
-import { Provider } from "react-redux";
-import Userhome from "./Components1/Userhome";
-import Home from "./Components1/Home";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { Store } from "lucide-react";
-import store from "./store/store";
-import React from "react";
+
+import { Outlet } from "react-router-dom";
+import Header from "./Components/Header";
+
 export default function App() {
+
+
   return (
-    <>
-      <BrowserRouter>
-      <Provider store={store}>
-      <Routes>
-          <Route path="/" element={<Home />} />
-        </Routes>
-      </Provider>
-      </BrowserRouter>
-    </>
+    <div className="w-full min-h-screen flex flex-col">
+      <div>
+        <Header/>
+      </div>
+      
+      <main className="flex-1">
+        <Outlet />
+      </main>
+
+      {/* You can add a footer here if needed */}
+    </div>
   );
+  // return (
+  //   <>
+  //     <BrowserRouter>
+  //     <Provider store={store}>
+  //     <Routes>
+  //         <Route path="/" element={<Homepage />} />
+  //       </Routes>
+  //     </Provider>
+  //     </BrowserRouter>
+  //   </>
+  // );
 }
