+++ frontends/src/Components/ui/Spin.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import './spin.css'
+
+function Spin() {
+  return (
+  
+    <div class="loader"></div>
+
+
+  )
+}
+
+export default Spin
