--- frontends/src/Components1/home.css
@@ -1,13 +1,0 @@
-.header{
-    position: relative;
-       background-image: url('../assets/snapchatt.png');
-       background-size: cover;
-       background-position: center;
-       transform: rotate(-20deg);
-       transform-origin: 60% 50%;
-       background-repeat: no-repeat;
-   }
-   .fake {
-   font-family: cursive;
-     box-shadow: 4px 4px rgb(235, 229, 67);
-   }
