--- frontends/src/Components1/ui/Spin.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import './spin.css'
-
-function Spin() {
-  return (
-  
-    <div class="loader"></div>
-
-
-  )
-}
-
-export default Spin
