+++ frontends/src/pages/Homepage.jsx
@@ -1,0 +1,19 @@
+  import React from 'react'
+  import Userhome from '../Components/Userhome';
+  import '../Components/CssFiles/home.css'
+
+
+
+  function Homepage() {
+    
+  
+
+  
+    return (
+      <div className='home w-screen h-screen overflow-x-hidden'>
+        <Userhome />
+      </div>
+    )
+  }
+
+  export default Homepage
